Change load logic to hourly
Motivation is that certain date directories contained
too many files and that local storage was running out of
space while staging before s3 load.

By breaking up the download and upload step to use hourly granularity
rather than date, this problem may be avoided.

It is notable that the the log of files uploaded remains
in date granularity. ie, the completed record file path
does not need to be changed.